requirements.txt
# Core
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.21
Werkzeug==2.3.7

# Database
psycopg2-binary==2.9.9

# Security
bcrypt==4.0.1
email-validator==2.0.0

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0

# PDF & Reports
reportlab==4.0.4
fpdf2==2.7.5

# Data processing
pandas==2.1.1
